fix(Menu.py): APIs return correct status code

+++ be/routes/Menu.py
@@ -1,4 +1,4 @@
-from flask import request, jsonify
+from flask import request, jsonify, make_response
 from flask.views import MethodView
 
 from .models.Menu import Menu
@@ -10,10 +10,14 @@ class MenuView(MethodView):
         if id is None:
             with app.app_context():
                 menu = db.session.query(Menu).all()
+                if len(menu) == 0:
+                    return make_response('Not found', 404)
             return jsonify([m.to_dict() for m in menu])
         else:
             with app.app_context():
                 menu = db.session.query(Menu).filter_by(id=id).first()
+                if menu is None:
+                    return make_response('Not found', 404)
             return jsonify(menu.to_dict())
 
     def post(self):
@@ -26,10 +30,11 @@ class MenuView(MethodView):
                 hide_price=data['hide_price'],
                 visibility=data['visibility'],
             )
+
             db.session.add(menu)
             db.session.commit()  # commit the transaction
             menu = db.session.query(Menu).filter_by(id=menu.id).first()
-            return jsonify(menu.to_dict())
+            return make_response(jsonify(menu.to_dict()), 201)
 
     def put(self, id: int):
         data = request.get_json()
@@ -37,6 +42,8 @@ class MenuView(MethodView):
         print(id)
         with app.app_context():
             menu = db.session.query(Menu).filter_by(id=id).first()
+            if menu is None:
+                return make_response('Not found', 404)
             menu.name = data['name']
             menu.description = data['description']
             menu.icon = data['icon']
@@ -49,6 +56,8 @@ class MenuView(MethodView):
     def delete(self, id: int):
         with app.app_context():
             menu = db.session.query(Menu).filter_by(id=id).first()
+            if menu is None:
+                return make_response('Not found', 404)
             db.session.delete(menu)
             db.session.commit()
-            return jsonify(menu.to_dict())
+            return make_response(jsonify({'result': True}), 204)
